Fix MNIST_SMALL

+++ train_and_test_cnn.py
@@ -2,6 +2,7 @@
 # train and test a convolutional neural network
 # ----------------------------------------------------
 
+from tensorflow.python.framework import dtypes
 import tensorflow as tf 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -66,7 +67,7 @@ def main():
 
 		complete_dataset = input_data.read_data_sets("MNIST_data/", one_hot=True)
 
-		small_training_dataset = DataSet(complete_dataset.train._images[:N], complete_dataset.train._labels[:N], reshape=False)
+		small_training_dataset = DataSet(complete_dataset.train._images[:N], complete_dataset.train._labels[:N], dtype=dtypes.uint8, reshape=False)
 
 		dataset = Datasets(train=small_training_dataset, validation = complete_dataset.validation, test=complete_dataset.test)
